Moved column to own module.
